Add a charset validator to the checkers.
CharSetValidator is a checker that will enforce rules for
commits and branches.

By default it will check your commits for non-UTF-8-valid text
and reject any such changes. There is also a fast-UTF-8 checker
that uses less resources than doing a full check. It's basically
a trade off between full validation (as in, can I actually decode
this properly) and just checking the 'unicode encoding' of the
strings.

Branches/tags is checked with a regexp and this can be configured
by the system administrator - from our experience we have found
that locking it to us-ascii lowercase is how we enforce fewer errors.

The branch check will only be enforced on new branches/tags.

Change-Id: I6779f5e83cc6b10d430710e4b7125b00b7a3b464