Update patch set 3
Patch Set 3:

(3 comments)

Sorry, must have been a mixup somewhere a long the lines... 

We're pretty sure that we built this and ran tests on it but unless there is issues with maven i can't explain it. Anyway, the issues identified below are fixed and it compiles yet again.

Patch-set: 3
Label: Verified=0