Add validator that checks for symbolic links
Project owner can now configure symbolic links. Commits with symbolic
links will be rejected.

For this check it was necessary to extend CommitUtils with a new method
which provides the possibility to add a callback method. This callback
method will be called, if the TreeWalk finds an entry that is different
to the entry of all parents.

Change-Id: I513c6ae10afdd61029a84a8eb0d55a7ae1c55538