Move functionality of PathValidator into util class
Every validator that wants to know which files changed in the current
push was forced to extend PathValidator.java. With the extraction of
this functionality into ChangesUtils.java, a validator doesn't have to
extend this class.

Also there is a method to get the set of paths that changed in this
commit.

Change-Id: I84975741725244603ae7150da89c0f985fbbd398