Update patch set 4
Patch Set 4:

> - can you add a test case for tar.gz? I am not a Java expert
 > (anymore) to judge whether getFileExtension("foo.tar.gz") returns
 > tar.gz or gz.

This is a good point. We need to check and also document that.

 > 
 > - As a user I would probably add gz as a blacklisted file
 > extension, but tar.gz seems to be a valid value as well.
 > 
 > - another interesting test case might be any combination of
 > blocked: .core  |  core
 > files      :  foo.core | core
 > 
 > - in addition it might make sense to allow .ext as well as ext
 > (without dot)
 > 
 > 
 > lastly I think the problems above could all be eliminated by having
 > the possibility to specify patterns to reject. this would make the
 > extensions just a subset of the possibilities:
 > .*\.core
 > core.*
 > .*\.xls.*


This may be a valid requirement just n that case this validator wouldn't
be an "Extension" validator but a filename pattern validator.

I would also like not to enforce usage of a regex. If someone wants to
block the "exe" extension that it should be as simple as writing:

  blockedFileExtension = exe


 > 
 > etc.

Patch-set: 4
Label: Code-Review=0