Update patch set 4
Patch Set 4:

> > the validation value might be introduced after a too long
 > filename
 > > has made it into the repository. I do not see a testcase that
 > > checks whether deleting such a file is still possible.
 > 
 > I think that the same applies for the blocked keywords. A keyword
 > which already made it in the repository may become blocked later.
 > Should it be allowed to modify such a file when the modification
 > doesn't touch the line with the blocked keyword?

In my opinion any change touching a file that is not complaint should be rejected. No matter whether it introduces the blocking criteria.
If the project owner defines such a criteria, he strictly does not want to have such content in the repository.
In most cases the issues are easy and quick to fix in a separate change.

Patch-set: 4