Add ContentTypeUtil to detect binary files
Some checks shouldn't run on binary files (e.g.  InvalidLineEndingCheck).
The InvalidLineEndingCheck already had a config parameter to define which
files are binary. But the BlockedKeywordCheck also needs to know which
files are binary. Because of this the ContentTypeUtil is introduced.

The config parameter ignoreFilesWhenCheckLineEndings is removed and a
new parameter binaryType is introduced to configure which content types
are considered binary.

To detect content type, we use Apache Tika [1]. For using Apache Tika a
new dependency [2] is needed.

It is not possible to use the already provided MimeUtil2 library from
Gerrit's plugin-api because MimeUtil2 [3] wrongly detects, for example,
the content type of project.config as application/octet-stream. Apache
Tika properly detects content type of the project.config as text/plain.
Also, it seems that MimeUtil2 is not further developed.

[1] https://tika.apache.org/
[2] org.apache.tika:tika-core:1.12
[3] http://mime-util.sourceforge.net/project-info.html

Change-Id: Ia2529f187b8362ad2f21bb5e9bd9d0f88eb775b5