Add ContentTypeUtil to detect binary files
Some checks shouldn't run on binary files (e. g.
InvalidLineEndingCheck). The InvalidLineEndingCheck already had a menu
entry to define which files are binary. But the BlockedKeywordCheck also
needs to know which files are binary. Because of this the
ContentTypeUtil is introduced. The ContentTypeUtil can detect if a file
is a binary or not.

The menu entry ignoreFilesWhenCheckLineEndings was removed, because it
is not longer needed.

To detect binary files ContentTypeUtil is using Apache Tika [1]. For
using Apache Tika a new dependency [2] is needed.

It is not possible to use the already existing module MimeUtil2Module,
because MimeUtil2 [3] does not support as many types as Apache Tika. And
it seems that MimeUtil2 is deprecated.

[1] https://tika.apache.org/
[2] org.apache.tika:tika-core:1.12
[3] http://mime-util.sourceforge.net/project-info.html

Change-Id: Ia2529f187b8362ad2f21bb5e9bd9d0f88eb775b5