Add ContentTypeUtil to detect binary files
Some checks shouldn't run on binary files (e. g.
InvalidLineEndingCheck). The InvalidLineEndingCheck already had a menu
entry to define which files are binary. But the BlockedKeywordCheck also
needs to know which files are binary. Because of this the
ContentTypeUtil is introduced. The ContentTypeUtil can detect if a file
is a binary or not.

The menu entry ignoreFilesWhenCheckLineEndings was removed, because it
is not longer needed.

Change-Id: Ia2529f187b8362ad2f21bb5e9bd9d0f88eb775b5