Give permission to override the rules
At times people need to be able to avoid the upload validation
rules to actually manage the repository structure or perform
specific operations.

E.g. A release manager should be able to upload some generated
artifacts and thus override the blockage of upload binary files
that is imposed to all other users.

Change-Id: I1e11288cd60690754ff4a38499dfcc4385190125