Allows exception to the rules
At times people need to be able to avoid the upload validation
rules to actually manage the repository structure or perform
specific operations.

E.g. A release manager should be able to upload some generated
artifacts and thus override the blockage of upload binary files
that is imposed to all other users.

By defining exception to the rules, it is possible to allow
individual users or groups to skip the rules for projects or branches.

Change-Id: I1e11288cd60690754ff4a38499dfcc4385190125