Fix blocking by keyword to apply only to changes
When editing existing files, the blocking of the keywords need to
apply only to the edited portion and not to the entire content.

Example: block all 'println' in the code needs to block commits
to a file that historically already contained existing println
statements. Without this fix, any modifications to the file would
have always been rejected.

Bug-Id: Issue 5353
Change-Id: Ifed4abfe2d3fb83f50fd35e921ff7fa766cb80ec