Remove Tika dependency for uploadvalidator
Tika was originally used over the MimeUtils2 library
because it was claimed that project.config was detected
as application/octet-stream. While project.config is in
fact detected as text/x-ini, other .config files are
detected as application/octet-stream.

Once change 103290 makes all .config files detected as
text/x-ini, we can remove the dependency on Tika introduced
in change 76268.

Change-Id: Ib9a64557b89058fc246a611c3e8a1be74eecf042