Update patch set 3
Patch Set 3:

It looks like the content type detection is used for two purposes:

1. to find binary files that should be skipped for some checks, like in BlockedKeyworkValidator and InvalidLineEndingValidator
2. to block certain content types, in ContentTypeValidator

For (1), I think jgit RawText's isBinary is a better fit. That way, unknown text based content types won't be misdetected as binary and made to skip the check.

For (2), I have no special opinions. MimeUtil2 seems a lot simpler than Tika, so this change is likely good for it.

Patch-set: 3
CC: Jonathan Nieder <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>