Use jgit RawText's isBinary in addition to MimeTypes
RawText's isBinary will allow us to detect whether or not
a file is text-based regardless of the content type, allowing
us to alleviate some misdetection based on bianry types.

As a backup, users will still be able to configure whether or not a file
is binary based on the content type.

Change-Id: I2726af4bf3235ee7a03cc06d88c9877f7081ae4b