Adds Email Whitelist Uploadvalidator
This adds the ability to define whitelists that Author and Committer
emails can be validated against when a push occurs.

The emails may be configured using specific emails, patterns, or
regular expressions. When enabled, all emails that do not match will
have push rejected.

Author and Committer will have seperate whitelists, but they follow
all the same rules when checking against their lists for validation.

Defining the whitelist is done on project level in the
project.config file of the project. Multiple entries can be defined
for both Author and Committer and usage the can be mixed. Usage being
specific email, pattern, or regular expression.

Change-Id: I27c9398994e1ee6bb1c6553a3144e8d998f2b020