Adds Change Email Uploadvalidator
This adds the ability to define patterns that will allow Author and
Committer emails that match to push changes. All others are rejected.
This check uses 'java.util.regex.Pattern'.

Author and Committer will have separate pattern lists, but they follow
all the same rules when checking against their lists for validation.

Defining the list of patterns is done on project level in the
project.config file of the project. Multiple entries can be defined
for both Author and Committer and usage the can be mixed. Usage being
specific email or regular expression.

Change-Id: I27c9398994e1ee6bb1c6553a3144e8d998f2b020