Fix file extension detecting for partial match in extension
If a short file extension like `.o` was configured to be blocked,
also files with an extension ending with an `o` like `.mako` were
blocked. This happened, because just the last characters of the
filename were checked.

Now a `.` is expected in front of the extension and thus enforcing
a comparison of the full file extension.

Change-Id: Ic480b72a7b677df5a05ec9cac17a90eade9aba1c