Avoid any uses or lookups of ffmpeg by Tika
I75d8b85d7 reintroduced a dependency on Tika to fix an issue. Switching
back to Tika introduces another issue, though. On startup, Tika reaches
out to the local system to find out whether ffmpeg is installed, which
it will then automatically use as parser. Depending on the environment
the uploadvalidator plugin is run, this lookup of ffmpeg on the local
system might not be permitted.

There are two options to avoid this lookup of ffmpeg:
1) Use a custom Tika config which excludes ExternalParser (which is
responsible for the ffmpeg lookup).
2) Directly use the default detector of Tika as we don't need the parser
functionality anyway.

In this change, we went with 2). For 1), we would have needed to
identify which exact configuration we're currently using. With 2),
we're also safe to not introduce a regression if we change the config
without further considerations in the future.

A third option would be to remove the Tika dependency but compared to
that, this change seems like a better option.

Change-Id: I76ec0c488f93e4a9c71f0527895acbc798c5a9ba