Replace tika by overviewproject mime-types to detect mime types
Use the org.overviewproject:mime-types:0.1.3 library [1] instead of
Apache Tika to detect mime-types. This library is a successor of the
original project mime-util [2] which seems to be unmaintained.
It supports the mime-types supported by shared-mime-info [3].

Compare to Apache Tika the overviewproject mime-types library has a much
smaller scope and only cares about detecting mime-types whereas Tika
provides a lot more functionality which is not needed in the
uploadvalidator plugin. The size of the mime-types library is only 58kB
compared to 708kB for the tika-core library.

[1] https://github.com/overview/mime-types
[2] http://sourceforge.net/p/mime-util/
[3] https://gitlab.freedesktop.org/xdg/shared-mime-info

Change-Id: Iecdae64b2039bd6717926e1b6dcb1f59b04a593e