Fix NPE in uploadvalidator
Change I4081375e033 migrated uploadvalidator to use DiffOperations. The
implementation uses two different sources:
1) CommitUtils#getChangedContent to get the changed files (regular,
executable) and their content.
2) DiffOperations (previously PatchListCache) to get the list of edits.

We loop on changed files using the file paths of [1], but we check the
edits from [2]. We need to check if the file path exists in
DiffOperation's output map. Sometimes DiffOperations skips some files,
for example if files that were due to rebase.

Bug: Google b/203216894
Change-Id: Ib6dce965de5a2e79b28e1a6f76c36c8049734020