Project Config Validation for Upload Validator Plugin
Adds new listener to validate changes to project.config. Email
restrictions will be checked for legal regex, illegal regex will cause
an error to be thrown and the commit to be rejected.

Validation functionality will be extended to other fields as well.
Email restrictions are currently not checked for legal regex when the
upload validator settings are changed in the UI.

Change-Id: Ied7fbe5b2d52ac40b04ac7f9cd5043e6064d855d