Project Config Validation for Upload Validator Plugin
Added new listener to validate changes on the uploadvalidator sub-
section of project.config. Email restrictions are checked for legal
regex, illegal regex will cause an error to be thrown and the commit
to be rejected.

Validation functionality will be extended to other fields as well.
Email restrictions are currently not checked for legal regex when the
upload validator settings are changed in the UI: This is a limitation
in gerrit core that will be fixed there.

Change-Id: Ied7fbe5b2d52ac40b04ac7f9cd5043e6064d855d