Use DiffOperationsForCommitValidation for listing modified files in validator
Using DiffOperations to list modified files from commit validators is
problematic as it may create Auto-Merge commits with a new
ObjectInserter instance, see change Id3630cfc6.

Instead we use the DiffOperationsForCommitValidation instance that is
provided in the CommitReceivedEvent. The provided
DiffOperationsForCommitValidation instance is guaranteed to use the same
ObjectInserter instance as the main thread, hence we avoid creating the
same Auto-Merge objects with 2 different (competing) ObjectInserter's.

BlockedKeywordValidator needs the file diffs of the modified files so
after loading the modified files via
DiffOperationsForCommitValidation.loadModifiedFilesAgainstParentIfNecessary
which returns ModifiedFile's we use
DiffOperations.getModifiedFileAgainstParent to load the file diffs.
Using DiffOperations for loading the file diffs is safe, as this doesn't
create Auto-Merge commits, but is a pure read operation.

To keep the behaviour of BlockedKeywordValidator unchanged we need to
skip modified files which are only touched due to rebased edits, as the
old code used DiffOperations.listModifiedFilesAgainstParent with
DiffOptions.DEFAULTS that does this.

This change follows the example of change I99b6eb6fc that did a similar
change for the code-owners plugin.

Bug: Google b/371217868
Change-Id: I7a53a9e2d53b63376d4504b645a792323b392638
Signed-off-by: Edwin Kempin <ekempin@google.com>