Add a commit-message overrides validation to UploadValidator
The new validator warns if the commit-message to-be-uploaded reverts the
commit-message to an older revision.

Unlike most of the validator tests, the added test is an integration test since
the validator deals with older patch sets, and the IT infrastructure mimics
this state better than the ValidatorTestCase one.

Change-Id: I7eae4d6c3aee0e797cf2995afcd39ea9cd940a0a
Release-Notes: skip
Google-Bug-Id: b/40015365