Initialize DuplicatePathnameValidator with full project hierarchy config
The docs specify that this plugin should read the full project's
hirearchy when parsing the config. This is done correctly at commit
validation time, however this is not true at plugin inizialization time.
This has the effect than if enabling rejectDuplicatePathnames in all
projects, the setting will be applied correctly when validating commits
however the UI for the project will show the 'Reject Duplicate
Pathnames' as off, rather than on, leading to confusion amongst users.

Bug: Issue 399587135
Change-Id: I8c0fc543fb97b19d928474beaccb2580807a3583